RD-2790 Split DrilldownButtons into a few files
Split a single file containing mulitple components into smaller files.
This way it should be easier to comprehend, extend, and modify.
