+++ widgets/common/src/deploymentsView/detailsPane/index.tsx
@@ -1,7 +1,7 @@
 import type { FunctionComponent } from 'react';
 
 import type { Deployment } from '../types';
-import DrilldownButtons, { DrilldownButtonsProps } from './DrilldownButtons';
+import DrilldownButtons, { DrilldownButtonsProps } from './drilldownButtons';
 import DetailsPaneHeader from './header';
 import DetailsPaneWidgets from './widgets';
 import type { SharedDeploymentsViewWidgetConfiguration } from '../configuration';
